GDI+ plugin Layout::Layout: suppress double→REAL conversion warning
This squashes the following compiler warning:

  C:\Program Files (x86)\Microsoft Visual Studio\2022\BuildTools\VC\Tools\MSVC\14.44.35207\include\memory(3630,56): warning C4244: 'argument': conversion from 'double' to 'Gdiplus::REAL', possible loss of data [C:\GitLab-Runner\builds\graphviz\graphviz\build\plugin\gdiplus\gvplugin_gdiplus.vcxproj]
    (compiling source file '../../../plugin/gdiplus/gvtextlayout_gdiplus.cpp')
        C:\Program Files (x86)\Microsoft Visual Studio\2022\BuildTools\VC\Tools\MSVC\14.44.35207\include\memory(3630,56):
        the template instantiation context (the oldest one first) is
            C:\GitLab-Runner\builds\graphviz\graphviz\plugin\gdiplus\gvtextlayout_gdiplus.cpp(61,15):
            see reference to function template instantiation 'std::unique_ptr<Gdiplus::Font,std::default_delete<Gdiplus::Font>> std::make_unique<Gdiplus::Font,const Gdiplus::FontFamily*,double&,0>(const Gdiplus::FontFamily *&&,double &)' being compiled

In practice, we do not expect the font size to be an out-of-range value.

Reported-by: MSVC 19.44.35223.0

+++ plugin/gdiplus/gvtextlayout_gdiplus.cpp
@@ -58,7 +58,7 @@ Layout::Layout(char *fontname, double fontsize, char* string)
 		font = std::make_unique<Font>(reference.hdc, &found_font);
 	}
 	else
-		font = std::make_unique<Font>(FontFamily::GenericSerif(), fontsize);
+		font = std::make_unique<Font>(FontFamily::GenericSerif(), static_cast<REAL>(fontsize));
 }
 
 void gdiplus_free_layout(void *layout)
